Commit 4: Added: Multiple Users; Accounts (Google) & Login Button; Rearranged Data Input & Associated it with Specific Users
Disregard .html ∆’s from lines 13 ~ 17 - copied in older code at one
point, the code in green was there from the beginning…

The addition of the ‘accounts’ functionality (based on ‘accounts-ui’ &
‘accounts-google’ packages) necessitated several ∆s to the ‘Users’
references as the accounts system established its own ‘user’ collection.

‘Users’ was ∆’ed to ‘ProteinData’; ‘allUsers’ was ∆’ed to
‘allProteinData’; and the ‘users’ collection was renamed to
‘protein_data’.

∆’ed the History insert to save the ‘userId’ instead of the history
entry/item id  - in effect adding a ‘userId’ key to each history entry.

Added the ‘userId’ ‘key’ as filters to the History.find calls.

With the addition of the Accounts function that allows new users to
signIn:

• Removed the √ for/initialize the User from the
  Meteor.startup function &

• Added it to new √ for existing data for the current user, &

• Included initializing a ‘protein_data’ collection doc for the
user if new.

Moved the ‘limit’ of 5 to the isServer section so that only 5 records
(not every one in a users data doc) would be sent to the client. Left
the sort on the client, but also added one on the server-side find to
help ensure that only the latest 5 records would be sent & would be
arranged LIFO in display.