Commit 5  Session Data
Set up Session.set & Session.get - in this case, to store text box
input by user between page refreshes - allows for ‘hot-fix’; i.e.
pushing fixes or new code to the browser without having to shut down
the app server & restart.

Placed the Session.set in the ‘userDetails’ ‘events’ code block where
the user’s input is accepted…

Placed the Session.get in the ‘userDetails’  ‘helpers’ code block to
fetch the amount from the session store which then gets passed to the
html template.

Placed a Handlebars template section in the .html ‘userDetails’
template block to capture the data element stored in the session.

Had to pull the text input element out of the {{#with user}} {{/with}}
block to disassociate the text box input from the ‘user’ context **not
really clear on why this was necessary… ?!?**