Commit 6: Add  Reactive 'Computation' via Meteor Dependency
A Dependency on a ‘reactive’ element, such as a template, will cause
the Dependency (‘Computation’) code to fire whenever the template data
(model) changes.

In this case, the app will log to the console who/when a user logs in &
logs out.

To do this, a ‘computation’ with an if:else is used inside a
‘Deps.autorun’ code block.

The Dependency  function will use the native ‘Meteor.user()’ function
to test if Meteor.user() is true (i.e. a user is logged in) or false &
whenever that value changes, (i.e., a user logs in or out), log the
appropriate  message to the console.

Meteor .user()