Commit 8: Adding Simple Permissions...
Added a subtract 100 action to the amount.

Then added a ‘ProteinData.deny’ function to check what the the result
of the subtract WOULD be, and allow it if it doesn’t take the total
below 0, or deny it if the subtraction would result in a negative
number.

In this case the default ‘insecure’ package has not yet been removed,
but in a normal app where it would be removed, this code would need to
take the inverse approach; i.e., from the ‘allow’ POV as opposed to
‘deny’.

Also added inline comments for the isSimulation code block.